refactor: utils, libs 폴더 구조를 변경합니다.
