+++ src/frontend/app/pages/landing/components/footer.tsx
@@ -1,6 +1,6 @@
 import { ArrowDownToLine } from 'lucide-react'
 
-import deviceHelper from '@/libs/device-helper'
+import deviceHelper from '@/utils/device-helper'
 
 function Bottom() {
   return (
