+++ src/frontend/app/pages/landing/components/home-hero.tsx
@@ -1,6 +1,6 @@
 import { ArrowDownToLine } from 'lucide-react'
 
-import deviceHelper from '@/libs/device-helper'
+import deviceHelper from '@/utils/device-helper'
 
 function HomeHero() {
   return (
