+++ src/frontend/src/components/select-box/index.tsx
@@ -1,7 +1,7 @@
 import { ChevronDownIcon } from 'lucide-react'
 import { Dispatch, ReactNode, SetStateAction, useRef, useState } from 'react'
 
-import { cn } from '@/utils/cn'
+import { cn } from '@/libs/cn'
 import useClickOutside from '@/hooks/use-click-outside'
 
 interface Props {
