+++ src/frontend/src/constants/discord-assets-path.ts
