+++ src/frontend/src/utils/device-helper.ts
